Add a bunch of test-only deps
The major important parts here are poltergeist, webmock, and Geocoder
test mod. poltergeist allows running js tests with phantomjs, webmock
ensures that tests don't depend on random outgoing HTTP requests, and we
add some test locations for Geocoder.

We need database_cleaner for js tests, since the browsing thread is
different from the test thread, so it doesn't see uncommitted
transactions. So for js tests, we use DatabaseCleaner's truncation
strategy, and the rest of the time, we use the faster transaction
strategy.

The phantomjs gem is just included to make it easier to get a working
phantomjs binary if you don't already have one on your PATH.